food_delivery_backend/requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
pyyaml
pymysql
